### Added * Added support for `login-path` to store your mysql db credentials in an encrypted file, and thus avoid `mysql: [Warning] Using a password on the command line interface can be insecure.` warnings * Added `LOCAL_DB_HOST` and `LOCAL_DB_PORT` to the `.env.sh` * Added `REMOTE_MYSQL_CMD` and `REMOTE_MYSQLDUMP_CMD` to the `.env.sh` * Added the `crontab-helper` file you can add to your `crontab` as a crib sheet * Added the `cache` table as one to delete/ignore
### Changed
* Made the scripts more tolerant of missing settings (perhaps due to old `.env.sh` files)
* Refactored common code out to `common/common_db.sh` and `common/common_env.sh`
* Updated the `README.md` with information about `fileMode`
* Updated the `README.md` with information about `login-path`

+++ scripts/pull_db.sh
@@ -7,7 +7,7 @@
 # @author    nystudio107
 # @copyright Copyright (c) 2017 nystudio107
 # @link      https://nystudio107.com/
-# @package   craft-permissions
+# @package   craft-scripts
 # @since     1.0.2
 # @license   MIT
 
@@ -16,37 +16,31 @@ if [[ ! -f ".env.sh" ]] ; then
     echo 'File ".env.sh" is missing, aborting.'
     exit
 fi
-
 source ".env.sh"
 
-# Tables to exclude from the dump
-EXCLUDED_TABLES=(
-            "assetindexdata"
-            "assettransformindex"
-            "cache"
-            "sessions"
-            "templatecaches"
-            "templatecachecriteria"
-            "templatecacheelements"
-            )
+# Make sure the `common_db.sh` exists
+if [[ ! -f "common/common_db.sh" ]] ; then
+    echo 'File "common/common_db.sh" is missing, aborting.'
+    exit
+fi
+source "common/common_db.sh"
 
-IGNORED_TABLES_STRING=""
-for TABLE in "${EXCLUDED_TABLES[@]}"
-do
-   IGNORED_TABLES_STRING+=" --ignore-table=${LOCAL_DB_NAME}.${GLOBAL_DB_TABLE_PREFIX}${TABLE}"
-done
+# Make sure the `common_env.sh` exists
+if [[ ! -f "common/common_env.sh" ]] ; then
+    echo 'File "common/common_env.sh" is missing, aborting.'
+    exit
+fi
+source "common/common_env.sh"
 
 # Temporary db dump path (remote & local)
 TMP_DB_PATH="/tmp/${REMOTE_DB_NAME}-db-dump-$(date '+%Y%m%d').sql"
 BACKUP_DB_PATH="/tmp/${LOCAL_DB_NAME}-db-backup-$(date '+%Y%m%d').sql"
 
 # Get the remote db dump
-REMOTE_DB_CREDS="--user=${REMOTE_DB_USER} --password=${REMOTE_DB_PASSWORD} --host=${REMOTE_DB_HOST} --port=${REMOTE_DB_PORT} ${REMOTE_DB_NAME}"
-ssh $REMOTE_SSH_LOGIN -p $REMOTE_SSH_PORT "mysqldump $REMOTE_DB_CREDS --single-transaction --no-data > '$TMP_DB_PATH' ; mysqldump $REMOTE_DB_CREDS --no-create-info $IGNORED_TABLES_STRING >> '$TMP_DB_PATH'"
+ssh $REMOTE_SSH_LOGIN -p $REMOTE_SSH_PORT "$REMOTE_MYSQLDUMP_CMD $REMOTE_DB_CREDS --single-transaction --no-data > '$TMP_DB_PATH' ; $REMOTE_MYSQLDUMP_CMD $REMOTE_DB_CREDS --no-create-info $IGNORED_TABLES_STRING >> '$TMP_DB_PATH'"
 scp -P $REMOTE_SSH_PORT -- $REMOTE_SSH_LOGIN:"$TMP_DB_PATH" "$TMP_DB_PATH"
 
 # Backup the local db
-LOCAL_DB_CREDS="--user=${LOCAL_DB_USER} --password=${LOCAL_DB_PASSWORD} ${LOCAL_DB_NAME}"
 $LOCAL_MYSQLDUMP_CMD $LOCAL_DB_CREDS --single-transaction --no-data > "$BACKUP_DB_PATH"
 $LOCAL_MYSQLDUMP_CMD $LOCAL_DB_CREDS --no-create-info $IGNORED_TABLES_STRING >> "$BACKUP_DB_PATH"
 echo "*** Backed up local database to $BACKUP_DB_PATH"
